LPS-32484 Simplify
